Create form for application creation
Django modelForm with all fields corrsponding to the Application model, except for created_at which is excluded from the modelForm, as date of creation is set to be generated automatically. Form widgets are also set for each field. In addition, for all names there were set validators(to check if they only contain letters) and for phone number field (to check if it only contains digits. In addition minimum length for the phone number has been set, so the form would not be submitted without the full number filled in.