Create form for creating a student
An inbuilt UserCreationForm for the CustomUser model. Includes the same fields as the Application form + username and student group fileds. Each field has it's limitations(min/max length) defined, as well as form widgets. The password fields are excluded, as they are to be automatically generated in the corresponding view function.  The save() function is overridden as besides saving the CustomUser data, it is necessary to save Student's data. For the created CustomUser in this function, a corresponding related Student object is created. In addition, in this function a user group is set to 'student' value and the password is to be set from the function's parameter.