Create view function for creating a new student
The function renders the template for the student creation form and processes it then. If there is a POST request - the data from the CreateStudentForm is processed and if it's valid - a user password is generated and passed to the the save method of that form, which then creates the user. In addition, user's username and email is extracted from the form and passed used to send the username and password to the newly created user. In case of successes or errors the teacher get's the corresponding messages.