Add login_required decorator and check user's rights for the ciew
 The login_required checks if the user is logged in. If user passes thes check, the view is allowed to be accessed, otherwise, the user is redirected to login page. The login url is specified in the settings.
In this view, the group_check function is not applied, as the users other from teach and superusers are allowed to access the lessons page (student). But in the application it is necessary to only allow a student to access the lessons of their group, without access to lessons of other students. Thus, in the view function it is checked if the current user is among the students who have these lessons' student group set as their student gorup, or if the current user is a teacher or a supeuser. If not, the user is redirected to login page