Add login_required decorator and check user's tights for the view
 The login_required checks if the user is logged in. If user passes this check, the view is allowed to be accessed, otherwise, the user is redirected to login page. The login url is specified in the settings.
In this view, the group_check function is not applied, as the users other from teacher and superusers are allowed to access the observations page (student). But in the application it is necessary to only allow a student to access observations of their own, without access to observations of other students. Thus, in the view function it is checked if the current user is the same user that the observations belong to, or if the current user is a teacher or a superuser. If not, the user is redirected to login page